Say Command instead of ComScr
`ComScr` is both too explicit and too limited
as to what a "command" may be.

The level of abstraction for commands is
shifting a bit, to be closer in spirit to the
POSIX shell, where a command is either a
call to an executable, a function call, or a
compound command.

FunTask merely adds another option: a script.